# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LiveForge HTTP API
  version: 1.1.0
  description: |
    Source-aligned contract for the management API, GB28181 API, WebRTC
    WHIP/WHEP signaling, and the default internal cluster relay signaling
    paths. Management responses use `{code,message,data?}`. GB28181 handlers
    return direct JSON objects or arrays. WebRTC signaling returns SDP or
    plain-text errors.

    Management and GB28181 routes accept either an opaque bearer token from
    `api.auth` or an authenticated `lf_session` console cookie when management
    credentials are configured. If no management tokens are configured,
    compatibility mode resolves an anonymous admin principal. Health is always
    public. A missing/invalid credential returns 401; an authenticated role
    without the operation's `x-liveforge-permission` returns 403. Enabled
    per-IP rate limiting can return 429 before a handler runs.

    Internal relay signaling paths are configurable through
    `cluster.rtp.signaling_path` and `cluster.gb28181.signaling_path`; this
    document shows their defaults. Node clients read credentials atomically
    for each request, preferring `api.auth.bearer_token`, then the first named
    admin token. See `docs/recipes/cluster-relay-operations.md` before exposing
    these internal paths.
servers:
  - url: http://127.0.0.1:8090
    description: Default local management listener
security:
  - bearerAuth: []
  - consoleSession: []
...
paths:
...
  /api/v1/recordings:
    get:
      tags: [recording]
      operationId: listRecordings
      summary: List recordings from configured storage
      x-liveforge-permission: recordings:read
      parameters:
        - {$ref: '#/components/parameters/ListLimit'}
        - {$ref: '#/components/parameters/ListOffset'}
        - {name: q, in: query, schema: {type: string}, description: Case-sensitive recording ID or stream-key substring.}
      responses:
        '200': {description: Recordings, content: {application/json: {schema: {$ref: '#/components/schemas/RecordingsResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/recordings/{recordingPath}:
    parameters:
      - {$ref: '#/components/parameters/RecordingPath'}
    get:
      tags: [recording]
      operationId: getRecording
      summary: Read recording metadata or explicitly play/download the full recording ID
      description: Without action, an existing exact recordingPath returns metadata. Set action=play or action=download to address that full ID unambiguously, including an ID whose final segment is play or download. Existing exact IDs take precedence over the backward-compatible suffix action forms. Media actions acquire the global connection budget before opening the recording and apply a 10-second write deadline; exhaustion returns 503. Only completed recordings are served; active or failed recordings return 409 with a JSON error and no media body.
      parameters:
        - name: action
          in: query
          required: false
          description: Explicit media action for the complete recordingPath.
          schema: {type: string, enum: [play, download]}
      x-liveforge-permission: recordings:read
      responses:
        '200': {description: Recording metadata or complete media, content: {application/json: {schema: {$ref: '#/components/schemas/RecordingResponse'}}, application/octet-stream: {schema: {type: string, format: binary}}, video/mp4: {schema: {type: string, format: binary}}, video/x-flv: {schema: {type: string, format: binary}}, video/mp2t: {schema: {type: string, format: binary}}}}
        '206': {description: Requested media byte range, content: {application/octet-stream: {schema: {type: string, format: binary}}, video/mp4: {schema: {type: string, format: binary}}, video/x-flv: {schema: {type: string, format: binary}}, video/mp2t: {schema: {type: string, format: binary}}}}
        '304': {description: Action response not modified}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/RecordingNotReady'}
        '416': {description: Requested action range is not satisfiable}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    delete:
      tags: [recording]
      operationId: deleteRecording
      summary: Delete a completed recording
      description: Always treats the complete recordingPath as the ID, even when its final segment is play or download. Local cleanup removes exact owned sidecars and metadata before the primary; a cleanup error leaves the primary available for retry.
      x-liveforge-permission: recordings:delete
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/recordings/{recordingPath}/download:
    parameters:
      - {$ref: '#/components/parameters/RecordingPath'}
    get:
      tags: [recording]
      operationId: downloadRecording
      summary: Download a completed recording with HTTP range support
      description: Backward-compatible suffix action used only when no exact recording ID includes the final /download segment. Use action=download on /api/v1/recordings/{recordingPath} for an unambiguous full-ID request. Only completed recordings are served; active or failed recordings return 409 with a JSON error and no media body. The media response acquires the global connection budget before opening the recording and applies a 10-second write deadline; exhaustion returns 503.
      x-liveforge-permission: recordings:read
      responses:
        '200': {description: Complete recording, content: {application/octet-stream: {schema: {type: string, format: binary}}}}
        '206': {description: Requested byte range, content: {application/octet-stream: {schema: {type: string, format: binary}}}}
        '304': {description: Not modified}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/RecordingNotReady'}
        '416': {description: Requested range is not satisfiable}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/recordings/{recordingPath}/play:
    parameters:
      - {$ref: '#/components/parameters/RecordingPath'}
    get:
      tags: [recording]
      operationId: playRecording
      summary: Stream a completed recording inline with HTTP range support
      description: Backward-compatible suffix action used only when no exact recording ID includes the final /play segment. Use action=play on /api/v1/recordings/{recordingPath} for an unambiguous full-ID request. Returns the recording with a media MIME type and Content-Disposition inline. The endpoint accepts HTTP Range requests for browser seeking; active or failed recordings return 409 with a JSON error and no media body. The media response acquires the global connection budget before opening the recording and applies a 10-second write deadline; exhaustion returns 503.
      x-liveforge-permission: recordings:read
      responses:
        '200': {description: Complete recording media, content: {video/mp4: {schema: {type: string, format: binary}}, video/x-flv: {schema: {type: string, format: binary}}, video/mp2t: {schema: {type: string, format: binary}}}}
        '206': {description: Requested byte range, content: {video/mp4: {schema: {type: string, format: binary}}, video/x-flv: {schema: {type: string, format: binary}}, video/mp2t: {schema: {type: string, format: binary}}}}
        '304': {description: Not modified}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/RecordingNotReady'}
        '416': {description: Requested range is not satisfiable}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque management token}
    consoleSession: {type: apiKey, in: cookie, name: lf_session}
    streamBearerAuth: {type: http, scheme: bearer, bearerFormat: publish or subscribe token}
    streamToken: {type: apiKey, in: query, name: token}
  parameters:
...
    ListLimit:
      name: limit
      in: query
      description: Enables bounded pagination; 1..500, defaults to 100 when only offset is supplied. Without both pagination parameters the legacy complete data shape is preserved. Filters apply first. Paginated JSON adds top-level pagination and X-Total-Count/X-Limit/X-Offset headers. Offset pages are live snapshots, not stable cursors across mutations.
      schema: {type: integer, minimum: 1, maximum: 500}
    ListOffset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
    RecordingPath: {name: recordingPath, in: path, required: true, description: URL-decoded recording ID relative to configured storage; traversal is rejected., schema: {type: string}}
...
  responses:
    Success: {description: Successful management operation, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
    BadRequest: {description: Invalid or missing input, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Unauthorized: {description: Missing or invalid management credential, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Forbidden: {description: Principal lacks the required permission, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Conflict: {description: Resource state conflicts with the operation, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    RecordingNotReady: {description: Recording is active or failed and has no downloadable media; the response is JSON and never contains the media body, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    RateLimited: {description: Per-IP request limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    InternalError: {description: Storage or server failure, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Unavailable: {description: Required module or runtime manager is disabled, closed, or unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties: {code: {type: integer, const: 0}, message: {type: string}, data: {}, pagination: {$ref: '#/components/schemas/Pagination'}}
    Pagination:
      type: object
      required: [limit, offset, total, has_more]
      properties: {limit: {type: integer}, offset: {type: integer}, total: {type: integer}, has_more: {type: boolean}}
    ApiError:
      type: object
      required: [code, message]
      properties: {code: {type: integer, minimum: 400}, message: {type: string}}
...
    Recording:
      type: object
      properties: {id: {type: string}, stream_key: {type: string}, format: {type: string, enum: [flv, fmp4, mp4, ts, hls]}, state: {type: string, enum: [active, completed, failed]}, size_bytes: {type: integer, format: int64}, started_at: {type: string, format: date-time}, completed_at: {type: string, format: date-time}, duration_sec: {type: number}, error: {type: string}}
...
    RecordingsResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Recording'}}}}]}
    RecordingResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {$ref: '#/components/schemas/Recording'}}}]}
...